requirements.txt
aiolimiter==1.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
cryptography==46.0.5
fastapi==0.135.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
packaging
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-telegram-bot==22.7
setuptools==80.10.2
socksio==1.0.0
starlette==1.0.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
wheel==0.46.3
